requirements.txt
torchvision~=0.17.0
transformers~=4.37.2
gradio~=4.19.0
numpy~=1.24.4
pillow~=10.2.0
tqdm~=4.66.2
requests~=2.31.0
grad-cam
scikit-image~=0.19.3
matplotlib~=3.7.4
pycocotools~=2.0.7
opencv-python~=4.7.0.72